+++ src/hexen/sv_save.cpp
@@ -3272,7 +3272,7 @@ static void CopyFile(char *source_name, char *dest_name)
     const int BUFFER_CHUNK_SIZE = 0x10000;
 
     uint8_t *buffer;
-    int file_length, file_remaining;
+    long file_length, file_remaining;
     FILE *read_handle, *write_handle;
     int buf_count, read_count, write_count;
